requirements.txt
Django

django-debug-toolbar
django-split-settings
django-extensions
django-bootstrap-pagination

Celery
django-haystack
djangorestframework

psycopg2
simplejson
Pillow
requests
elasticsearch
pdfminer.six
PyPDF2
Werkzeug